+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Ignore forwarded `Like` activities.
 - Set 10 minute timeout on background job that processes incoming activities.
 - Use "warn" log level for delivery errors.
+- Don't allow read-only users to manage subscriptions.
 
 ### Fixed